# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Taco Hunt API
  version: 1.0.0
  description: Contrato HTTP del API Taco Hunt. Las rutas de escritura requieren autenticación.
servers:
  - url: http://localhost:3001
paths:
...
  /v1/admin/queue:
    get:
      operationId: getModerationQueue
      summary: Obtiene la cola privada de moderación
      security: [{ bearerAuth: [] }]
      parameters:
        - {
            in: query,
            name: kind,
            required: true,
            schema: { type: string, enum: [spots, tacos, reports] },
          }
      responses:
        "200":
          description: Elementos pendientes de moderación
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { type: object, additionalProperties: true }
        "400": { $ref: "#/components/responses/ApiError" }
        "401": { $ref: "#/components/responses/ApiError" }
        "403": { $ref: "#/components/responses/ApiError" }
  /v1/admin/audit:
    get:
      operationId: listModerationAudit
      summary: Consulta el historial privado de acciones administrativas
      security: [{ bearerAuth: [] }]
      parameters:
        - {
            in: query,
            name: limit,
            schema: { type: integer, minimum: 1, maximum: 100, default: 50 },
          }
        - {
            in: query,
            name: before,
            description: Cursor opaco devuelto por la página anterior,
            schema: { type: string, maxLength: 512 },
          }
      responses:
        "200":
          description: Acciones recientes y cursor anterior
          content:
            application/json:
              schema:
                type: object
                required: [items, nextBefore]
                properties:
                  items:
                    { type: array, items: { $ref: "#/components/schemas/ModerationAuditEntry" } }
                  nextBefore: { type: [string, "null"], maxLength: 512 }
        "400": { $ref: "#/components/responses/ApiError" }
        "401": { $ref: "#/components/responses/ApiError" }
        "403": { $ref: "#/components/responses/ApiError" }
  /v1/admin/spot-proposals/{id}/approve:
    post:
      operationId: approveSpotProposal
      summary: Aprueba un puesto propuesto
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Puesto aprobado }
        "400": { $ref: "#/components/responses/ApiError" }
        "401": { $ref: "#/components/responses/ApiError" }
        "403": { $ref: "#/components/responses/ApiError" }
        "404": { $ref: "#/components/responses/ApiError" }
  /v1/admin/spot-proposals/{id}/reject:
    post:
      operationId: rejectSpotProposal
      summary: Rechaza un puesto propuesto y conserva un motivo interno
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ModerationReason" }
      responses:
        "200": { description: Puesto rechazado }
        "400": { $ref: "#/components/responses/ApiError" }
        "401": { $ref: "#/components/responses/ApiError" }
        "403": { $ref: "#/components/responses/ApiError" }
        "404": { $ref: "#/components/responses/ApiError" }
  /v1/admin/taco-proposals/{id}/approve:
    post:
      operationId: approveTacoProposal
      summary: Aprueba un tipo de taco propuesto
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Tipo de taco aprobado }
        "400": { $ref: "#/components/responses/ApiError" }
        "401": { $ref: "#/components/responses/ApiError" }
        "403": { $ref: "#/components/responses/ApiError" }
        "404": { $ref: "#/components/responses/ApiError" }
  /v1/admin/taco-proposals/{id}/reject:
    post:
      operationId: rejectTacoProposal
      summary: Rechaza un tipo de taco propuesto y conserva un motivo interno
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ModerationReason" }
      responses:
        "200": { description: Tipo de taco rechazado }
        "400": { $ref: "#/components/responses/ApiError" }
        "401": { $ref: "#/components/responses/ApiError" }
        "403": { $ref: "#/components/responses/ApiError" }
        "404": { $ref: "#/components/responses/ApiError" }
  /v1/admin/reviews/{id}/hide:
    post:
      operationId: hideReview
      summary: Oculta una reseña de las respuestas públicas
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ModerationReasonOptional" }
      responses:
        "200": { description: Reseña oculta }
        "400": { $ref: "#/components/responses/ApiError" }
        "401": { $ref: "#/components/responses/ApiError" }
        "403": { $ref: "#/components/responses/ApiError" }
        "404": { $ref: "#/components/responses/ApiError" }
  /v1/admin/reviews/{id}/unhide:
    post:
      operationId: unhideReview
      summary: Vuelve a mostrar una reseña
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ModerationReasonOptional" }
      responses:
        "200": { description: Reseña visible }
        "400": { $ref: "#/components/responses/ApiError" }
        "401": { $ref: "#/components/responses/ApiError" }
        "403": { $ref: "#/components/responses/ApiError" }
        "404": { $ref: "#/components/responses/ApiError" }
  /v1/admin/reports/{id}/close:
    post:
      operationId: closeReport
      summary: Cierra un reporte después de su revisión
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ModerationReasonOptional" }
      responses:
        "200": { description: Reporte cerrado }
        "400": { $ref: "#/components/responses/ApiError" }
        "401": { $ref: "#/components/responses/ApiError" }
        "403": { $ref: "#/components/responses/ApiError" }
        "404": { $ref: "#/components/responses/ApiError" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ApiError:
      description: Error con identificador de solicitud
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              {
                type: string,
                enum:
                  [
                    VALIDATION_ERROR,
                    UNAUTHORIZED,
                    FORBIDDEN,
                    NOT_FOUND,
                    CONFLICT,
                    RATE_LIMITED,
                    SERVICE_UNAVAILABLE,
                    INTERNAL_ERROR,
                  ],
              }
            message: { type: string }
            requestId: { type: string }
            details: { type: object, additionalProperties: true }
...
    ModerationReason:
      type: object
      required: [reason]
      additionalProperties: false
      properties:
        reason: { type: string, minLength: 1, maxLength: 500, pattern: '\S' }
    ModerationReasonOptional:
      type: object
      additionalProperties: false
      properties:
        reason:
          {
            type: string,
            minLength: 1,
            maxLength: 500,
            pattern: '\S',
            default: "Acción de moderación",
          }
    ModerationAuditEntry:
      type: object
      required: [id, moderatorId, targetType, targetId, action, createdAt, reason]
      properties:
        id: { type: string, format: uuid }
        moderatorId: { type: [string, "null"], format: uuid }
        targetType: { type: string }
        targetId: { type: string, format: uuid }
        action: { type: string }
        createdAt: { type: string, format: date-time }
        reason: { type: [string, "null"], maxLength: 500 }